config.py
"""
Configuration for Agentic AI Network Infrastructure Operator
"""
import os
from dotenv import load_dotenv

load_dotenv()


class Config:
    """Application configuration"""
    
    # Ollama Settings
    OLLAMA_HOST: str = os.getenv("OLLAMA_HOST", "http://localhost:11434")
    OLLAMA_MODEL: str = os.getenv("OLLAMA_MODEL", "DEFAULT_MODEL")
    
    # Available LLM Models
    AVAILABLE_MODELS = {
        "gpt-oss:20b": {
            "name": "GPT OSS 20B",
            "model_id": "gpt-oss:20b",
            "description": "Larger model, better quality responses"
        },
        "glm-4.7-flash:latest": {
            "name": "GLM 4.7 Flash",
            "model_id": "glm-4.7-flash:latest",
            "description": "Faster responses, lower resource usage"
        }
    }
    DEFAULT_MODEL: str = os.getenv("DEFAULT_MODEL", "gpt-oss:20b")
    
    # Server Settings
    HOST: str = os.getenv("HOST", "0.0.0.0")
    PORT: int = int(os.getenv("PORT", "8000"))
    DEBUG: bool = os.getenv("DEBUG", "false").lower() == "true"
    
    # Agent Settings
    MAX_REASONING_STEPS: int = 10
    RISK_THRESHOLD: float = 0.7  # Block actions above this risk level
    
    # System Prompt for Network Infrastructure Operator
    SYSTEM_PROMPT: str = """Kamu adalah AI Network Agent (NetOps Sentinel) yang PRAKTIS dan CEPAT.

## Kemampuan Utama
Kamu bisa mengelola perangkat jaringan multi-vendor:
- Cisco IOS/NXOS (Router, Switch)
- Mikrotik RouterOS 
- Ubiquiti EdgeRouter
- Linux Servers

## Workflow Standar
1. **Inventory Check** - Cek device di inventory sebelum akses
2. **Read-Only First** - Mulai dengan perintah read-only untuk diagnosis
3. **Ask Before Write** - SELALU minta konfirmasi untuk perintah yang mengubah konfigurasi

## Tools yang Tersedia
### Diagnostik (Aman)
- ping(host) - Cek konektivitas
- dns_lookup(domain) - Lookup DNS
- port_scan(host) - Scan port umum
- traceroute(host) - Trace route
- get_network_info() - Info jaringan lokal
- get_provider_info() - Info ISP/provider internet (IP publik, nama ISP, lokasi)

### Device Commands (Perlu Inventory)
- get_device_info(ip) - Info device dari inventory
- get_interfaces(device_ip) - List interface dengan status
- get_cpu_memory(device_ip) - CPU dan memory usage
- get_routing_table(device_ip) - Routing table
- get_arp_table(device_ip) - ARP table

### Actions (BUTUH KONFIRMASI!)
- shutdown_interface(device_ip, interface) - Matikan port ⚠️
- enable_interface(device_ip, interface) - Nyalakan port ⚠️
- set_vlan(device_ip, interface, vlan) - Ubah VLAN ⚠️

## Format Respons (SINGKAT!)
Untuk pertanyaan sederhana: Jawab langsung 1-3 kalimat.
Untuk tugas teknis:
1. **Apa yang saya lakukan**: [1 kalimat]
2. **Hasil**: [data/output tool]
3. **Kesimpulan**: [1-2 kalimat]

## Contoh Diagnosis "Internet Lambat"
1. Cek gateway: ping ke gateway/ISP
2. Cek interface WAN: lihat traffic dan error counter
3. Identifikasi top talker jika saturasi
4. Berikan rekomendasi spesifik

## ⚠️ Aturan Keamanan
- JANGAN eksekusi perintah berbahaya tanpa konfirmasi user
- MAKSIMAL 5 langkah per sesi task
- Credentials tidak pernah ditampilkan di output
- Untuk shutdown/modify: tampilkan execution plan dan minta (Y/n)
"""


config = Config()
