requirements.txt
# Core Framework
ollama
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv
httpx
rich
jinja2
websockets
psutil>=5.9.0

# LangChain + LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.3.0
langgraph>=0.2.0

# Multi-Vendor Network Drivers
netmiko>=4.2.0
napalm>=4.1.0
paramiko>=3.4.0

# CLI Output Parsing
textfsm>=1.1.3
ntc-templates>=4.0.0

# Inventory & IPAM
pynetbox>=7.3.0
aiosqlite>=0.19.0

# Vector Database for RAG
chromadb>=0.4.0

# DeepSeek LLM Integration
openai>=1.12.0

